messing with main and eeprom to include some stuff for setting IP to be able to eventually send an ARP message
added connect command and for now got rid of clear so that there's no terminal interference with the same command beginning witth the letter 'C'.

Also set default IP to 192.168.1.111 to be used for connecting and stored in EEPROM for now... need to add stuff so user can change if wanted.

Mainly just testing to get an ARP message through